move load_files and chdir to utils.py
