+++ tests/utils.py
@@ -1,0 +1,29 @@
+import contextlib
+import json
+import os
+import pathlib
+import tempfile
+import unittest
+from typing import *
+
+
+@contextlib.contextmanager
+def load_files(files: Dict[str, bytes]) -> Iterator[pathlib.Path]:
+    with tempfile.TemporaryDirectory() as tempdir_:
+        tempdir = pathlib.Path(tempdir_)
+        for relpath, data in files.items():
+            path = tempdir / relpath
+            path.parent.mkdir(parents=True, exist_ok=True)
+            with open(str(path), "wb") as fh:
+                fh.write(data)
+        yield tempdir
+
+
+@contextlib.contextmanager
+def chdir(path: pathlib.Path) -> Iterator[None]:
+    cwd = os.getcwd()
+    try:
+        os.chdir(str(path))
+        yield
+    finally:
+        os.chdir(cwd)
